requirements.txt
asgiref==3.5.2
attrs==21.2.0
Automat==20.2.0
azure-core==1.26.4
azure-storage-blob==12.16.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto3==1.26.115
botocore==1.29.115
cachetools==5.3.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==22.4.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.3
django-crispy-forms==1.12.0
django-storages==1.13.2
dropbox==11.36.0
filelock==3.8.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
isodate==0.6.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pexpect==4.8.0
Pillow==9.4.0
pipenv==2022.11.11
platformdirs==2.5.4
ply==3.11
protobuf==4.22.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.4
sqlparse==0.4.3
ssh-import-id==5.11
stone==3.3.1
systemd-python==234
Twisted==22.1.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.16.7
virtualenv-clone==0.5.7
wadllib==1.3.6
WALinuxAgent==2.2.46
zipp==1.0.0
zope.interface==5.4.0
